src/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
PyYAML==6.0.1
requests==2.32.3
schedule==1.2.2
setuptools==69.5.1
soupsieve==2.5
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.43.0
